requirements.txt
fastapi==0.95.2
uvicorn==0.22.0
SQLAlchemy==1.4.53
pydantic==1.10.7
python-multipart==0.0.6
pdfplumber>=0.7,<0.12
python-docx==0.8.11
pandas==2.2.0
sentence-transformers==2.2.2
scikit-learn==1.3.2
cachetools==5.3.1
jinja2==3.1.2
